% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aaer_dates}
\alias{aaer_dates}
\title{AAER dates from SEC}
\format{
A tibble with 2,920 rows and 4 variables:
\describe{
\item{aaer_num}{AAER number}
\item{aaer_date}{Date}
\item{aaer_desc}{Description}
\item{year}{Year of AAER}
}
}
\usage{
aaer_dates
}
\description{
A data set containing dates and descriptions for AAERs.
}
\keyword{datasets}
