% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aaer_firm_year}
\alias{aaer_firm_year}
\title{AAERs from Bao et al. (2020)}
\format{
A tibble with 415 rows and 4 variables:
\describe{
\item{p_aaer}{AAER identifier}
\item{gkvey}{GVKEY (firm identifier)}
\item{min_year}{First affected year}
\item{max_year}{Last affected year}
}
}
\usage{
aaer_firm_year
}
\description{
A data set containing AAER firms-years used in Bao et al. (2020).
}
\keyword{datasets}
