% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apple_events}
\alias{apple_events}
\title{Dates for Apple Events}
\format{
A tibble with 47 rows and 3 variables:
\describe{
\item{event}{Description of event}
\item{event_date}{First date of event}
\item{end_event_date}{Last date of event}
...
}
}
\source{
\url{https://en.wikipedia.org/wiki/List_of_Apple_Inc._media_events}
}
\usage{
apple_events
}
\description{
A dataset containing the dates of Apple media events
since 2005.
}
\keyword{datasets}
