% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{auc}
\alias{auc}
\title{Area under curve}
\source{
\url{https://blog.mbq.me/augh-roc/}

\url{https://stackoverflow.com/questions/4903092/calculate-auc-in-r}
}
\usage{
auc(scores, response)
}
\arguments{
\item{scores}{Probability that response is true or 1.}

\item{response}{Responses coded as logical or {0, 1}.}
}
\value{
vector including AUC
}
\description{
A function returning AUC.
}
