% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aus_bank_funds}
\alias{aus_bank_funds}
\title{Australian bank fundamental data}
\format{
A tibble with 283 rows and 7 variables:
\describe{
\item{gvkey}{GVKEY (firm identifier)}
\item{datadate}{Fiscal year-end}
\item{at}{Total assets}
\item{ib}{Income before extraordinary items}
\item{xi}{Extraordinary items}
\item{do}{Income from discontinued operations}
}
}
\usage{
aus_bank_funds
}
\description{
A data set containing fundamental financial information for Australian banks.
}
\keyword{datasets}
