% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aus_bank_rets}
\alias{aus_bank_rets}
\title{Australian bank stock market data}
\format{
A tibble with 3,053 rows and 4 variables:
\describe{
\item{gkvey}{GVKEY (firm identifier)}
\item{datadate}{Last trading date of month}
\item{ret}{Stock return for month}
\item{mkt_cap}{Market capitalization on datadate}
}
}
\usage{
aus_bank_rets
}
\description{
A data set containing fundamental financial information for Australian banks.
}
\keyword{datasets}
