% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aus_banks}
\alias{aus_banks}
\title{Australian banks}
\format{
A tibble with 10 rows and 3 variables:
\describe{
\item{gkvey}{GVKEY (firm identifier)}
\item{ticker}{Stock exchange ticker}
\item{co_name}{Bank name}
}
}
\usage{
aus_banks
}
\description{
A data set containing identifying information for 10 Australian banks.
}
\keyword{datasets}
