% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallfar_nonpara_sthao.R
\name{boot_farr_fit.emp}
\alias{boot_farr_fit.emp}
\title{Non-parametric estimation of the far from bootstrap samples}
\usage{
boot_farr_fit.emp(x, z, rp, B = 100)
}
\arguments{
\item{x}{the variable of interest in the counterfactual world.}

\item{z}{the variable of interest in the factual world.}

\item{rp}{the return periods for which the far is to be estimated.}

\item{B}{the number of bootstrap samples to draw.}
}
\value{
An object of class \code{("boot_farrfit.emp", "boot_farrfit", "farrfit")}.
It is a matrix where each columm contains the far estimated for the returns periods \code{rp}
for a given bootstrap sample.
}
\description{
\code{boot_farr_fit.emp} returns an object of class \code{("boot_farrfit.emp", "boot_farrfit", "farrfit")}
which contains the estimations of the far for each bootstrap sample and
for different return periods \code{rp}
}
\details{
This function returns bootstrap non-parametric estimations of far, the fraction of attributable risk for records,
as defined in Naveau et al (2018).The far is estimate from the bootstrap samples of the data x and z
that are obtained by resampling bootstrap. 
The first bootstrap sample is obtained from the original dataset of x and z.

For the full reference, see : \emph{Naveau, P., Ribes, A., Zwiers, F., Hannart, A., Tuel, A., & Yiou, P.
Revising return periods for record events in a climate event attribution context.
J. Clim., 2018.}, \url{https://doi.org/10.1175/JCLI-D-16-0752.1}
}
\examples{
 library(evd)
 
 muF <-  1; xiF <- .15; sigmaF <-  1.412538 #  cst^(-xiF) # .05^(-xi1);
 # asymptotic limit for the far in this case with a Frechet distributiom
 boundFrechet <- 1 - sigmaF^(-1/xiF)
 # sample size
 size <- 100
 # level=.9
 set.seed(4)
 z = rgev(size, loc = (sigmaF), scale = xiF * sigmaF, shape = xiF)
 x = rgev(length(z), loc=(1), scale = xiF, shape=xiF)
 
 rp = seq(from = 2, to = 30, length = 200)
 
 # Resampling bootstrap for the non-parametrc estimation of the far
 boot_farr.emp <- boot_farr_fit.emp(x = x, z = z, rp = rp, B = 10)
 print(boot_farr.emp)
 confint(boot_farr.emp)
 
 ylim <- range(boundFrechet, boot_farr.emp)
 plot(boot_farr.emp, ylim = ylim, main = "boot far non-parametric")
 # Theoretical far for in this case (Z = sigmaF * X  with X ~ Frechet)
 lines(rp, frechet_farr(r = rp, sigma = sigmaF, xi = xiF), col = "red", lty = 2)
 abline(h = boundFrechet, col = "red", lty = 2)
}
