% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{camp_attendance}
\alias{camp_attendance}
\title{Camp attendance}
\format{
A tibble with 1,000 rows and 2 variables:
\describe{
\item{id}{Student identifier}
\item{camp}{Indicator for student attendance at camp}
}
}
\usage{
camp_attendance
}
\description{
A simulated data set related to camp attendance.
}
\keyword{datasets}
