% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{confusion_stats}
\alias{confusion_stats}
\title{Confusion statistics.}
\usage{
confusion_stats(scores, response, predicted = NULL, k = NULL)
}
\arguments{
\item{scores}{Probability that response is true or 1.}

\item{response}{Responses coded as logical or 0-or-1.}

\item{predicted}{Predicted value coded as 0-or-1.}

\item{k}{Percentage to classify as TRUE or 1.}
}
\value{
vector including sensitivity and precision
}
\description{
A function returning sensitivity and precision.
}
