% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxOakes.R
\name{CoxOakes}
\alias{CoxOakes}
\title{Test for exponentiality of Cox and Oakes}
\usage{
CoxOakes(x)
}
\arguments{
\item{x}{a numeric vector of positive values.}
}
\value{
A list containing the following components:
\describe{
  \item{statistic}{the value of the test statistic.}
  \item{p.value}{the p-value of the test.}
}
}
\description{
\code{CoxOakes} returns results of the test of exponentiallity of Cox and Oakes
}
\details{
This function performs the test of exponentiallity of Cox and Oakes.
For more details, see: Henze and Meintanis (2005, Metrika)

For the full reference, see : \emph{Naveau, P., Ribes, A., Zwiers, F., Hannart, A., Tuel, A., & Yiou, P.
Revising return periods for record events in a climate event attribution context.
J. Clim., 2018.}, \url{https://doi.org/10.1175/JCLI-D-16-0752.1}
}
\examples{
set.seed(1)
x <- rexp(1000)
CoxOakes(x)
}
