% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_pg.R
\name{df_to_pg}
\alias{df_to_pg}
\title{Convert a data frame to a remote data frame}
\usage{
df_to_pg(df, conn, read_only = TRUE)
}
\arguments{
\item{df}{data frame}

\item{conn}{connection to a PostgreSQL database}

\item{read_only}{Indicate that connection is read-only, so no compute() is possible}
}
\value{
tbl_sql
}
\description{
Convert a data frame to a remote PostgreSQL data frame
even with a read-only connection.
}
