% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallfar_para_sthao.R
\name{ecdf}
\alias{ecdf}
\title{Generic ecdf function}
\usage{
ecdf(x, ...)
}
\arguments{
\item{x}{the object to compute the ecdf from.}

\item{...}{additional arguments if necessary.}
}
\description{
Generic ecdf function
}
