% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallfar_nonpara_sthao.R
\name{estim_farr.emp}
\alias{estim_farr.emp}
\title{Non-parametric estimation of the far}
\usage{
estim_farr.emp(x, z, rp)
}
\arguments{
\item{x}{the variable of interest in the counterfactual world.}

\item{z}{the variable of interest in the factual world.}

\item{rp}{the return periods for which the far is to be estimated.}
}
\value{
An object of class \code{("farrfit_emp", "farrfit")}. It is a list containing the following
elements:
\describe{
  \item{rp}{the return periods for which the far is estimated.}
  \item{farr_hat}{the estimate of the far for each return period \code{rp}}
  \item{sigma_farr_hat}{the standard deviation of the estimator of the far
  assuming asymptotic gaussianity}.
}
}
\description{
\code{estim_farr.emp} returns an object of class  \code{("farrfit_emp", "farrfit")} which contains
the results of the non-parametric estimation of the far.
}
\details{
This function returns an non-parametric estimate of the far, the fraction of attributable risk for records,
as defined in Naveau et al (2018).

For the full reference, see : \emph{Naveau, P., Ribes, A., Zwiers, F., Hannart, A., Tuel, A., & Yiou, P.
Revising return periods for record events in a climate event attribution context.
J. Clim., 2018.}, \url{https://doi.org/10.1175/JCLI-D-16-0752.1}
}
\examples{
 library(evd)

 muF <-  1; xiF <- .15; sigmaF <-  1.412538 #  cst^(-xiF) # .05^(-xi1);
 # asymptotic limit for the farr in this case with a Frechet distributiom
 boundFrechet <- frechet_lim(sigma = sigmaF, xi = xiF)
 # sample size
 size <- 100
 # level=.9
 set.seed(4)
 z = rgev(size, loc = (sigmaF), scale = xiF * sigmaF, shape = xiF)
 x = rgev(length(z), loc=(1), scale = xiF, shape=xiF)

 rp = seq(from = 2, to = 30, length = 200)
 # non-parametric estimation of far
 farr_fit.emp <- estim_farr.emp(x = x, z = z, rp = rp)
 print(farr_fit.emp)
 ylim <- range(boundFrechet, farr_fit.emp$farr_hat)
 plot(farr_fit.emp, ylim = ylim, main = "far empirical")
 # Theoretical for in this case (Z = sigmaF * X  with X ~ Frechet)
 lines(rp, frechet_farr(r = rp, sigma = sigmaF, xi = xiF), col = "red", lty = 2)
 abline(h = boundFrechet, col = "red", lty = 2)
}
