% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallfar_para_sthao.R
\name{estim_theta.wexp}
\alias{estim_theta.wexp}
\alias{hist.thetafit_wexp}
\alias{qqplot.thetafit_wexp}
\alias{ecdf.thetafit_wexp}
\title{Parametric estimation of theta under H0: W ~ exp(theta)}
\usage{
estim_theta.wexp(x, z)

\method{hist}{thetafit_wexp}(x, ...)

\method{qqplot}{thetafit_wexp}(x, ...)

\method{ecdf}{thetafit_wexp}(x, ...)
}
\arguments{
\item{x}{the variable of interest in the counterfactual world.}

\item{z}{the variable of interest in the factual world.}

\item{...}{additional arguments for the plot.}
}
\value{
An object of class \code{("thetafit_wexp", "thetafit")}. It is a list containing the following
elements:
\describe{
  \item{theta_hat}{the estimate of theta}
  \item{sigma_theta_hat}{the standard deviation of the estimator of theta
  assuming asymptotic gaussianity and obtained via the delta-method}
  \item{W}{the estimate of W}
  \item{co_test}{the result of the Cox and Oakes test
  to test whether W follows an exponential distribution}
}
}
\description{
\code{estim_theta.wexp} returns an object of class  \code{("thetafit_wexp", "thetafit")} which contains
the results of the estimation of theta and of the fit of W ~ exp(theta).
}
\details{
This function returns an estimate of theta as defined in Naveau et al (2018) that is used
to estimate the far, the fraction of attributable risk for records (with \code{\link{estim_farr.wexp}}).
This estimation is made assuming that W = - log(G(Z)) follows an
exponentional distribution: W ~ exp(theta). G denotes the Cumulative Distribution Function of
the counterfactual variable X.

For the full reference, see : \emph{Naveau, P., Ribes, A., Zwiers, F., Hannart, A., Tuel, A., & Yiou, P.
Revising return periods for record events in a climate event attribution context.
J. Clim., 2018.}, \url{https://doi.org/10.1175/JCLI-D-16-0752.1}
}
\examples{
 library(evd)

 muF <-  1; xiF <- .15; sigmaF <-  1.412538 #  cst^(-xiF) # .05^(-xi1);
 # asymptotic limit for the far in this case with a Frechet distributiom
 boundFrechet <- frechet_lim(sigma = sigmaF, xi = xiF)
 # sample size
 size <- 100
 # level=.9
 set.seed(4)
 z = rgev(size, loc = (sigmaF), scale = xiF * sigmaF, shape = xiF)
 x = rgev(length(z), loc=(1), scale = xiF, shape=xiF)

 rp = seq(from = 2, to = 30, length = 200)
 # parametric estimation of far with exponential distribution for theta
 theta_fit <- estim_theta.wexp(x = x, z = z)
 # Check fit for W ~ exp(theta)
 hist(theta_fit)
 ecdf(theta_fit)
 qqplot(theta_fit)

 # Estimate the far
 farr_fit.exp <- estim_farr.wexp(theta_hat = theta_fit$theta_hat,
  sigma_theta_hat = theta_fit$sigma_theta_hat,
   rp = rp)
 print(farr_fit.exp)
 ylim <- range(boundFrechet, farr_fit.exp$farr_hat)
 plot(farr_fit.exp, ylim = ylim, main = "far exponential")
 # Theoretical for in this case (Z = sigmaF * X  with X ~ Frechet)
lines(rp, frechet_farr(r = rp, sigma = sigmaF, xi = xiF), col = "red", lty = 2)
 abline(h = boundFrechet, col = "red", lty = 2)
}
