% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farr.R
\docType{package}
\name{farr}
\alias{farr}
\alias{farr-package}
\title{farr: A package to compute the fraction of attributable risk for record events}
\description{
The \code{farr} package provides a set of functions to estimate
the fraction of attributable risk for record events, see for instance
\code{\link{estim_farr.emp}} and \code{\link{estim_farr.wexp}}.
}
\details{
For a full description of the method, see:
\emph{VNaveau, P., Ribes, A., Zwiers, F., Hannart, A., Tuel, A., & Yiou, P.
Revising return periods for record events in a climate event attribution context.
J. Clim., 2018.},
\url{https://doi.org/10.1175/JCLI-D-16-0752.1},
}
