% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_deciles.R
\name{form_deciles}
\alias{form_deciles}
\title{Form deciles}
\usage{
form_deciles(x)
}
\arguments{
\item{x}{A vector for which deciles are to be calculated.}
}
\value{
vector
}
\description{
Calculate deciles for a variable.
}
\examples{
library(farr)
library(dplyr, warn.conflicts = FALSE)

df <-
    tibble(x = rnorm(100)) \%>\%
    mutate(dec_x = form_deciles(x))
df
}
