% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annc_dates.R
\name{get_annc_dates}
\alias{get_annc_dates}
\title{Produce a table mapping announcements to trading dates}
\usage{
get_annc_dates(conn)
}
\arguments{
\item{conn}{connection to a PostgreSQL database}
}
\value{
tbl_df
}
\description{
Produce a table mapping announcements to trading dates
}
