% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ff_ind.R
\name{get_ff_ind}
\alias{get_ff_ind}
\title{Fetch Fama-French industry grouping.}
\usage{
get_ff_ind(ind, refresh_cache = FALSE)
}
\arguments{
\item{ind}{Fama-French industry grouping (e.g., 11, 48)}

\item{refresh_cache}{If \code{TRUE}, re-download the underlying zip file even if
a cached copy is available.}
}
\value{
tbl_df
}
\description{
Fetch Fama-French industry grouping from Ken French's website.
}
\examples{
## Not run:
get_ff_ind(5)
## End(Not run)
}
