% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_me_breakpoints.R
\name{get_me_breakpoints}
\alias{get_me_breakpoints}
\title{Create a table of with cut-offs for size portfolios}
\usage{
get_me_breakpoints()
}
\value{
tbl_df
}
\description{
Create a table of with cut-offs for size portfolios
}
