% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_size_rets_monthly.R
\name{get_size_rets_monthly}
\alias{get_size_rets_monthly}
\title{Create a table of monthly returns for size portfolios}
\usage{
get_size_rets_monthly()
}
\value{
tbl_df
}
\description{
Create a table of monthly returns for size portfolios
}
