% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_scores.R
\name{get_test_scores}
\alias{get_test_scores}
\title{A function returning data on test_scores.}
\usage{
get_test_scores(
  effect_size = 15,
  n_students = 1000L,
  n_grades = 4L,
  include_unobservables = FALSE
)
}
\arguments{
\item{effect_size}{Effect of attending camp on subsequent test scores.}

\item{n_students}{Number of students in simulated data set.}

\item{n_grades}{Number of grades in simulated data set.}

\item{include_unobservables}{Include talent in returned data (TRUE or FALSE)}
}
\value{
tbl_df
}
\description{
A function returning simulated data on test_scores.
}
