% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trading_dates.R
\name{get_trading_dates}
\alias{get_trading_dates}
\title{Produce a table mapping dates on CRSP to "trading days"}
\usage{
get_trading_dates(conn)
}
\arguments{
\item{conn}{connection to a PostgreSQL database}
}
\value{
tbl_df
}
\description{
Produce a table mapping dates on CRSP to "trading days".
Returned table has two columns: date, a trading date on CRSP;
td, a sequence of integers ordered by date
}
