% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{idd_dates}
\alias{idd_dates}
\title{Dates for Inevitable Disclosure Doctrine (IDD)}
\format{
A tibble with 24 rows and 3 variables:
\describe{
\item{state}{Two-letter state abbreviation}
\item{idd_date}{Date of precedent-setting legal case}
\item{idd_type}{Either "Adopt" or "Reject"}
}
}
\source{
\url{https://doi.org/10.1016/j.jfineco.2018.02.008}
}
\usage{
idd_dates
}
\description{
Dates of precedent-setting legal cases adopting or reject the
Inevitable Disclosure Doctrine (IDD) by state.
}
\keyword{datasets}
