% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iliev_2010}
\alias{iliev_2010}
\title{Data on public float}
\format{
A tibble with 7,214 and 9 variables:
\describe{
\item{gvkey}{Compustat firm identifier (GVKEY)}
\item{fyear}{Fiscal year}
\item{fdate}{Date of end of fiscal year}
\item{pfdate}{Date for public float value}
\item{pfyear}{Year for public float value}
\item{publicfloat}{Public float in $ million}
\item{mr}{Indicator for filing of a management report}
\item{af}{Indicator for accelerator filer}
\item{cik}{SEC firm identifier (CIK)}
}
}
\source{
\doi{10.1111/j.1540-6261.2010.01564.x}
}
\usage{
iliev_2010
}
\description{
Data on public float of listed companies from Iliev (2010).
}
\keyword{datasets}
