% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jar_fraud_features}
\alias{jar_fraud_features}
\title{Compustat features from Bao et al. (2020)}
\format{
A tibble with 146,045 rows and 30 variables:
\describe{
\item{fyear}{ }
\item{gvkey}{GVKEY (firm identifier)}
\item{act}{ }
\item{ap}{ }
\item{at}{ }
\item{ceq}{ }
\item{che}{ }
\item{cogs}{ }
\item{csho}{ }
\item{dlc}{ }
\item{dltis}{ }
\item{dltt}{ }
\item{dp}{ }
\item{ib}{ }
\item{invt}{ }
\item{ivao}{ }
\item{ivst}{ }
\item{lct}{ }
\item{lt}{ }
\item{ni}{ }
\item{ppegt}{ }
\item{pstk}{ }
\item{re}{ }
\item{rect}{ }
\item{sale}{ }
\item{sstk}{ }
\item{txp}{ }
\item{txt}{ }
\item{xint}{ }
\item{prcc_f}{ }
}
}
\usage{
jar_fraud_features
}
\description{
A data set containing features used in Bao et al. (2020).
}
\keyword{datasets}
