% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_parquet.R
\name{load_parquet}
\alias{load_parquet}
\title{Function to load parquet file into database.}
\usage{
load_parquet(conn, schema, table_name, data_dir = Sys.getenv("DATA_DIR"))
}
\arguments{
\item{conn}{DuckDB connection}

\item{schema}{Database schema for table}

\item{table_name}{Name of table to be loaded}

\item{data_dir}{Directory for data repository}
}
\value{
Remote data frame in conn
}
\description{
Function to read data from a parquet file data_dir/schema/table_name.parquet
into a table in the DuckDB database at conn.
}
