% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{ndcg}
\alias{ndcg}
\title{Calculate metric metric: NDCG at k}
\usage{
ndcg(scores, response, k = 0.01)
}
\arguments{
\item{scores}{Probability that response is true or 1.}

\item{response}{Responses coded as logical or {0, 1}.}

\item{k}{Percentage to classify as TRUE or 1.}
}
\value{
vector including sensitivity and precision
}
\description{
A function returning NDCG at k metric.
}
