% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_to_parquet.R
\name{pg_to_parquet}
\alias{pg_to_parquet}
\title{Save WRDS table as parquet file.}
\usage{
pg_to_parquet(table_name, schema, data_dir = Sys.getenv("DATA_DIR"))
}
\arguments{
\item{table_name}{Name of table on WRDS}

\item{schema}{Database schema for table}

\item{data_dir}{Directory for data repository}
}
\value{
Number of rows created
}
\description{
Function to get data from a table on the WRDS PostgreSQL server and
save to local parquet file using DuckDB.
}
