% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallfar_para_sthao.R
\name{qqplot}
\alias{qqplot}
\title{Generic qqplot function}
\usage{
qqplot(x, ...)
}
\arguments{
\item{x}{the object to compute the qqplot from.}

\item{...}{additional arguments if necessary.}
}
\description{
Generic qqplot function
}
