% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{roc}
\alias{roc}
\title{A function returning data for a ROC plot.}
\usage{
roc(scores, response)
}
\arguments{
\item{scores}{Probability that response is true or 1.}

\item{response}{Responses coded as logical or 0-or-1.}
}
\value{
tbl_df
}
\description{
A function returning data for a ROC plot.
}
