% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rusboost.R
\name{rus}
\alias{rus}
\title{Random under-sampling function
Function to create temporary training dataset using distribution implied}
\usage{
rus(y_train, weights, ir = 1)
}
\arguments{
\item{y_train}{df on the target variable.}

\item{weights}{Weights to be used in sampling data.}

\item{ir}{Imbalance ratio. Specifies how many times the under-sampled majority instances are over minority instances.}
}
\value{
vector
}
\description{
Random under-sampling function
Function to create temporary training dataset using distribution implied
}
