% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sho_firm_years}
\alias{sho_firm_years}
\title{Firm-years for replication of Fang, Huang and Karpoff (2016)}
\format{
A tibble with 60,272 rows × 2 variables.
\describe{
\item{gvkey}{GVKEY (firm identifier)}
\item{datadate}{Fiscal year-end}
}
}
\usage{
sho_firm_years
}
\description{
A data set containing the GVKEYs and
datadates for firm-years used in Fang, Huang and Karpoff (2016).
}
\keyword{datasets}
