% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sho_pilot}
\alias{sho_pilot}
\title{Treatment indicators for SHO pilot firms}
\format{
A tibble with 3,030 rows × 4 variables.
\describe{
\item{ticker}{Ticker}
\item{gvkey}{GVKEY (firm identifier)}
\item{permno}{PERMNO (CRSP security identifier)}
\item{pilot}{SHO pilot program treatment indicator}
}
}
\usage{
sho_pilot
}
\description{
A data set containing the tickers, GVKEYs, and
treatment indicator for SHO pilot program.
}
\keyword{datasets}
