% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sho_r3000}
\alias{sho_r3000}
\title{Russell 3000 stocks at time of SEC Reg SHO sample formation}
\format{
A tibble with 3000 rows × 2 variables.
\describe{
\item{russell_ticker}{Ticker}
\item{russell_name}{Company name}
}
}
\source{
\doi{10.1111/jofi.12369}
}
\usage{
sho_r3000
}
\description{
A data set containing the tickers and
company names for Russell 3000 at time SEC created the pilot sample.
Data are created from sample supplied by FHK.
}
\keyword{datasets}
