% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sho_r3000_gvkeys}
\alias{sho_r3000_gvkeys}
\title{Russell 3000 sample used by SEC with GVKEYs}
\format{
A tibble with 2,951 rows × 3 variables.
\describe{
\item{ticker}{Ticker}
\item{permno}{PERMNO (CRSP security identifier)}
\item{gvkey}{GVKEY (Compustat firm identifier)}
\item{pilot}{Indicator for stock being part of Reg SHO pilot program}
}
}
\source{
\url{https://iangow.github.io/far_book/natural-revisited.html#the-sho-pilot-sample}
}
\usage{
sho_r3000_gvkeys
}
\description{
A data set containing the tickers, PERMNOs, GVKEYs, and
treatment assignments for Russell 3000 sample used by SEC.
}
\keyword{datasets}
