% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sho_tickers}
\alias{sho_tickers}
\title{Tickers of pilot firms for Reg SHO.}
\format{
A tibble with 986 rows × 2 variables.
\describe{
\item{ticker}{Ticker}
\item{co_name}{Company name}
}
}
\source{
\url{https://www.sec.gov/rules/other/34-50104.htm}
}
\usage{
sho_tickers
}
\description{
A data set containing the tickers and
company names for pilot firms from Reg SHO pilot.
Data are scraped from the SEC's own website.
}
\keyword{datasets}
