% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_time.R
\name{system_time}
\alias{system_time}
\title{Version of \code{system.time()} that works with assignment}
\usage{
system_time(expr)
}
\arguments{
\item{expr}{Valid R expression to be timed, evaluated and returned}
}
\value{
Result of evaluating \code{expr}
}
\description{
Print CPU (and other) times that \code{expr} used, return value of \code{expr}.
}
