% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zhang_2007_events}
\alias{zhang_2007_events}
\title{Event dates from Zhang (2007)}
\format{
A tibble with 30 rows × 3 variables.
\describe{
\item{event}{Identifier for the event}
\item{date}{Date of event}
\item{event_desc}{Description of the event}
}
}
\source{
\doi{10.1016/j.jacceco.2007.02.002}
}
\usage{
zhang_2007_events
}
\description{
A data set containing the event dates used in Zhang (2007).
Data obtained from Panel of Table of Zhang (2007).
If an event spans multiple dates, then a row is
included for each date.
}
\keyword{datasets}
