% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zhang_2007_windows}
\alias{zhang_2007_windows}
\title{Event windows from Zhang (2007)}
\format{
A tibble with 17 rows × 3 variables.
\describe{
\item{event}{Identifier for the event}
\item{beg_date}{First date of event window}
\item{end_date}{Last date of event window}
}
}
\source{
\url{https://doi.org/10.1016/j.jacceco.2007.02.002}
}
\usage{
zhang_2007_windows
}
\description{
A data set containing the event windows used in Zhang (2007).
Data obtained from Panel of Table of Zhang (2007).
}
\keyword{datasets}
