% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bloomfield_2021}
\alias{bloomfield_2021}
\title{Firm-years in RDD analysis of Bloomfield (2021).}
\format{
A tibble with 1,855 rows and 2 variables:
\describe{
\item{fyear}{Fiscal year}
\item{permco}{CRSP firm identifier (PERMCO)}
}
}
\usage{
bloomfield_2021
}
\description{
Firm-years in RDD analysis of Bloomfield (2021).
}
\keyword{datasets}
