% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{by_tag_year}
\alias{by_tag_year}
\title{Tags on StackOverflow}
\format{
A tibble with 40,518 rows and 4 variables:
\describe{
\item{year}{Year}
\item{tag}{Tag}
\item{number}{Number of questions with tag during year}
\item{year_total}{Total number of questions with tag during year}
}
}
\usage{
by_tag_year
}
\description{
A data set containing data on tagged questions on StackOverflow
}
\keyword{datasets}
