% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxOakes.R
\name{CoxOakes}
\alias{CoxOakes}
\title{Test for exponentiality of Cox and Oakes}
\usage{
CoxOakes(x)
}
\arguments{
\item{x}{a numeric vector of positive values.}
}
\value{
A list containing the following components:
\describe{
  \item{statistic}{the value of the test statistic.}
  \item{p.value}{the p-value of the test.}
}
}
\description{
\code{CoxOakes} returns results of the test of exponentiallity of Cox and Oakes
}
\details{
This function performs the test of exponentiallity of Cox and Oakes.
For more details, see: Henze and Meintanis (2005, Metrika)
}
\examples{
set.seed(1)
x <- rexp(1000)
CoxOakes(x)
}
