% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_event_dates.R
\name{get_event_dates}
\alias{get_event_dates}
\title{Produce a table mapping announcements to trading dates}
\usage{
get_event_dates(
  data,
  conn,
  permno = "permno",
  event_date = "event_date",
  win_start = 0,
  win_end = 0,
  end_event_date = NULL
)
}
\arguments{
\item{data}{data frame containing data on events}

\item{conn}{connection to a PostgreSQL database}

\item{permno}{string representing column containing PERMNOs for events}

\item{event_date}{string representing column containing dates for events}

\item{win_start}{integer representing start of trading window (e.g., -1)}

\item{win_end}{integer representing start of trading window (e.g., 1)}

\item{end_event_date}{string representing column containing ending dates for events}
}
\value{
tbl_df
}
\description{
Produce a table of event dates for linking with CRSP.
See \code{vignette("wrds-conn", package = "farr")} for more on using this function.
}
\examples{
## Not run:
library(DBI)
library(dplyr, warn.conflicts = FALSE)
pg <- dbConnect(RPostgres::Postgres())
events <- tibble(permno = c(14593L, 10107L),
                 event_date = as.Date(c("2019-01-31", "2019-01-31")))
get_event_dates(events, pg, win_start = -3, win_end = + 3)
## End(Not run)
}
