% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ff_ind.R
\name{get_ff_ind}
\alias{get_ff_ind}
\title{Fetch Fama-French industry grouping.}
\usage{
get_ff_ind(ind)
}
\arguments{
\item{ind}{Fama-French industry grouping (e.g., 11, 48)}
}
\value{
tbl_df
}
\description{
Fetch Fama-French industry grouping from Ken French's website.
}
