% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_me_breakpoints.R
\name{get_me_breakpoints}
\alias{get_me_breakpoints}
\title{Create a table of with cut-offs for size portfolios}
\usage{
get_me_breakpoints(keep_max = FALSE)
}
\arguments{
\item{keep_max}{Set to \code{TRUE} to keep upper-bound of highest decile.
Default is \code{FALSE}, which will replace upper bound with \code{Inf}.}
}
\description{
Create a table of with cut-offs for size portfolios
}
\examples{
library(dplyr, warn.conflicts = FALSE)
get_me_breakpoints() \%>\% filter(month == '2022-04-01')
}
