% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gvkey_ciks}
\alias{gvkey_ciks}
\title{GVKEY-CIK links}
\format{
A tibble with 76,632 rows and 5 variables:
\describe{
\item{gvkey}{GVKEY (Compustat firm identifier)}
\item{iid}{Issue ID}
\item{cik}{CIK (SEC firm identifier)}
\item{first_date}{First link date}
\item{last_date}{Last link date}
}
}
\usage{
gvkey_ciks
}
\description{
Link table from GVKEYs to CIKs
}
\keyword{datasets}
