% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iliev_2010}
\alias{iliev_2010}
\title{Data on public float.}
\format{
A tibble with 7,213 and 6 variables:
\describe{
\item{gvkey}{Compustat firm identifier (GVKEY)}
\item{fyear}{Fiscal year}
\item{fdate}{Date of end of fiscal year}
\item{pfdate}{Date for public float value}
\item{pfyear}{Year for public float value}
\item{publicfloat}{Public float in $ million}
}
}
\usage{
iliev_2010
}
\description{
Data on public float of listed companies from Iliev (2010).
}
\keyword{datasets}
