% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{llz_2018}
\alias{llz_2018}
\title{GVKEYs used in Li, Lin and Zhang (2018).}
\format{
A tibble with 5,830 rows and 1 variable:
\describe{
\item{gvkey}{GVKEY}
}
}
\source{
\doi{10.1111/1475-679X.12187}
}
\usage{
llz_2018
}
\description{
GVKEYs used in Li, Lin and Zhang (2018).
}
\keyword{datasets}
