% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rusboost.R
\name{rus}
\alias{rus}
\title{Random under-sampling function}
\usage{
rus(y_train, ir = 1)
}
\arguments{
\item{y_train}{df on the target variable.}

\item{ir}{Imbalance ratio. Specifies how many times the under-sampled majority instances are over minority instances.}
}
\value{
vector
}
\description{
Function to create temporary training dataset using distribution implied
by w.
}
\details{
Following MATLAB, function samples observations of the minority class with
replacement and observations of the majority class without replacement.
}
