% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rusboost.R
\name{rusboost}
\alias{rusboost}
\title{RUSBoost for two-class problems}
\usage{
rusboost(formula, df, size, ir = 1, learn_rate = 1, rus = TRUE, control)
}
\arguments{
\item{formula}{A formula specify predictors and target variable. Target variable should be a factor of 0 and 1. Predictors can be either numerical and categorical.}

\item{df}{A df frame used for training the model, i.e. training set.}

\item{size}{Ensemble size, i.e. number of weak learners in the ensemble model}

\item{ir}{Imbalance ratio. Specifies how many times the under-sampled majority instances are over minority instances.}

\item{learn_rate}{Default of 1.}

\item{rus}{TRUE for random undersampling; FALSE for AdaBoost with full sample}

\item{control}{Control object passed onto rpart function.}
}
\value{
rusboost object
}
\description{
RUSBoost for two-class problems
}
