% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_scores}
\alias{test_scores}
\title{Test scores}
\format{
A tibble with 47 rows and 3 variables:
\describe{
\item{id}{Student identifier}
\item{grade}{School grade at time of test}
\item{post}{Indicator for being in grade 10 or 11}
\item{treat}{Indicator for student attending camp after grade 9}
\item{score}{Test score}
}
}
\source{
\url{https://en.wikipedia.org/wiki/List_of_Apple_Inc._media_events}
}
\usage{
test_scores
}
\description{
A simulated data set of test scores.
}
\keyword{datasets}
